fix + lint

+++ src/modules/DepositionGeant4/SensitiveScintillatorActionG4.hpp
@@ -47,14 +47,14 @@ namespace allpix {
         /**
          * @brief Get total number of hits deposited in the sensitive device bound to this action
          */
-        unsigned int getTotalScintillatorHits();
+        unsigned long getTotalScintillatorHits();
 
         /**
          * @brief Get the number of hits deposited in the sensitive device for this event only.
          * @warning The correct number is only available after dispatching the message, before it refers to the previous
          * event.
          */
-        unsigned int getScintillatorHits();
+        unsigned long getScintillatorHits();
 
         /**
          * @brief Get the name of the sensitive device bound to this action
@@ -85,8 +85,8 @@ namespace allpix {
         std::mt19937_64 random_generator_;
 
         // Statistics of total and per-event registered hits
-        unsigned int total_scint_hits_{};
-        unsigned int scint_hits_{};
+        unsigned long total_scint_hits_{};
+        unsigned long scint_hits_{};
 
         // Set of deposited charges in this event!!!!!!!!!!!!1!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!111
         std::vector<DepositedCharge> deposits_;
